+++ changelog.md
@@ -1,9 +1,20 @@
-## 2.1.3
-2024/july
+# 2024/july
+
+### 2.1.5
+by Paulo Martins
+dromedario.etc update
+	- fix Admin Logo
+	- getting put data on ApiControl
+	- fixing update from basic api crud
+	- fix Join Object 
+	- improved `ToArray` model function
+	- `CacheClear` function on `MagratheaApiControl`
+	- automatic cache on CRUD requests
+	- cache fix
+	- delete log file from admin
+
+### 2.1.3
 by Paulo Martins
 	- delete objects from magrathea_objects.conf
 	- API fallback
 	- move authentication endpoints to parent ApiAuthentication class
-
-
-